ks test - add type casting

+++ library/ml_defs.py
@@ -251,7 +251,7 @@ def ks_test(y_true,y_pred):
     a histogram
     """
     # return K.max(K.abs(K.cumsum(y_true,axis=1)-K.cumsum(y_pred,axis=1)))
-    return K.mean(K.square(K.cumsum(y_true,axis=1)-K.cumsum(y_pred,axis=1)),axis=1)
+    return K.mean(K.square(K.cumsum(K.cast(y_true,'float32'),axis=1)-K.cumsum(y_pred,axis=1)),axis=1)
 
 def poisson_nll(y_true,y_pred):
     """
